services/scorelab-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-cloud-aiplatform